Cope with newer builds of OpenSSL.
Newer builds are leaving out SHA-0 support. Handle this by including a
basic from-scratch implementation coded directly from FIPS180-1 (which
does describe the changes between SHA and SHA-1). Since the difference
was trivial a basic SHA-1 implemenatation is included as well. A basic
test is included too.

OpenSSL 1.1.0 makes the DH data structure opaque with accessors.
Furthermore, the file handle *_fp functions are not always built either.
Address these problems by rewriting all the Cyan-style DH file reading
and writing code to use BIO and to use the DH accessors. Also, work
around OpenSSL mistakenly requiring a public key for DH, see also
https://github.com/openssl/openssl/pull/4384

Fix up all the OpenSSL-related build code. The autoconf macro was not
properly handling the move of functionality from libssl to libcrypto.
It now also separately checks for SHA and RC4 (because the latter is
surely next to go). Add configure flags for each of these and fix up
all the #includes and #ifdefs. For test programs use the in-tree
SHA/SHA-1 unconditionally (it's simpler to do this), presuming the
OpenSSL one is better (though it is not too critical in the actual
server either).